fix: stay selected left menu when cluster namespace changed
