+++ src/redux/dashboard/slice.ts
@@ -16,6 +16,16 @@ export const dashboardSlice = createSlice({
     },
     setDashboardMenuList: (state: Draft<DashboardState>, action: PayloadAction<Array<DashboardMenu>>) => {
       state.ui.menuList = action.payload;
+      if (
+        action.payload.every(item => item.key !== state.ui.activeMenu.key) &&
+        action.payload.every(item =>
+          item.children
+            ? item.children.filter(i => i.resourceCount).every(i => i.key !== state.ui.activeMenu.key)
+            : true
+        )
+      ) {
+        state.ui.activeMenu = {key: 'Overview', label: 'Overview'};
+      }
     },
     setDashboardSelectedResourceId: (state: Draft<DashboardState>, action: PayloadAction<string | undefined>) => {
       state.tableDrawer.selectedResourceId = action.payload;
@@ -28,7 +38,6 @@ export const dashboardSlice = createSlice({
   extraReducers: builder => {
     builder.addCase(startClusterConnection.fulfilled, state => {
       state.tableDrawer.selectedResourceId = undefined;
-      state.ui.activeMenu = {key: 'Overview', label: 'Overview'};
     });
   },
 });
